requirements.txt
pandas>=2.0.0
requests==2.32.3
beautifulsoup4==4.12.3
sqlalchemy>=2.0.0
pytrends==4.9.2
lxml>=4.9.0
loguru==0.7.2

Django==5.1
djangorestframework==3.16.0
python-decouple==3.8
djangorestframework-simplejwt>=5.2.0
drf-spectacular>=0.27.0

psycopg2-binary>=2.9.0
numpy>=1.24.0
python-dateutil>=2.8.0
openpyxl>=3.1.0
xlrd>=2.0.0

PyGithub>=1.58.0
google-api-python-client>=2.88.0
python-dotenv==1.0.1

loguru==0.7.2

pytest>=7.0.0
reportlab>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
